:wrench: add AccessLogInterceptor to grpc server config
