+++ internal/servers/server/option.go
@@ -21,6 +21,7 @@ import (
 	"crypto/tls"
 	"fmt"
 	"net/http"
+	"strings"
 	"time"
 
 	"github.com/vdaas/vald/internal/errgroup"
@@ -28,6 +29,7 @@ import (
 	"github.com/vdaas/vald/internal/log"
 	"github.com/vdaas/vald/internal/net"
 	"github.com/vdaas/vald/internal/net/grpc"
+	"github.com/vdaas/vald/internal/net/grpc/interceptor/server/logging"
 	"github.com/vdaas/vald/internal/net/http/rest"
 	"github.com/vdaas/vald/internal/timeutil"
 )
@@ -420,8 +422,18 @@ func WithGRPCHeaderTableSize(size int) Option {
 	}
 }
 
-func WithGRPCInterceptors(name ...string) Option {
+func WithGRPCInterceptors(names ...string) Option {
 	return func(s *server) {
-		// s.grpc.opts = append(s.grpc.opts, grpc.UnaryInterceptor(uint32(size)))
+		for _, name := range names {
+			switch strings.ToLower(name) {
+			case "accessloginterceptor", "accesslog":
+				s.grpc.opts = append(
+					s.grpc.opts,
+					grpc.ChainUnaryInterceptor(logging.AccessLogInterceptor()),
+					grpc.ChainStreamInterceptor(logging.AccessLogStreamInterceptor()),
+				)
+			default:
+			}
+		}
 	}
 }
