+++ pkg/agent/core/ngt/usecase/agentd.go
@@ -80,14 +80,8 @@ func New(cfg *config.Data) (r runner.Runner, err error) {
 			vald.RegisterValdServer(srv, g)
 		}),
 		server.WithGRPCOption(
-			grpc.ChainUnaryInterceptor(
-				grpc.RecoverInterceptor(),
-				grpc.AccessLogInterceptor(),
-			),
-			grpc.ChainStreamInterceptor(
-				grpc.RecoverStreamInterceptor(),
-				grpc.AccessLogStreamInterceptor(),
-			),
+			grpc.ChainUnaryInterceptor(grpc.RecoverInterceptor()),
+			grpc.ChainStreamInterceptor(grpc.RecoverStreamInterceptor()),
 		),
 		server.WithPreStartFunc(func() error {
 			return nil
